pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=0.9.9",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.2.0",
    "langgraph-cli",
    "langchain-openai>=0.3.33",
]
